copied (with adaptations) overwrite/integrity check logic to first dl function

+++ app/satmo/download.py
@@ -9,7 +9,7 @@ from pprint import pprint
 from .utils import parse_file_name, make_file_path
 from .errors import HttpResourceNotAvailable
 
-def download_file(url, write_dir):
+def download_file(url, write_dir, overwrite = False, check_integrity = False):
     """Generic file download function
 
     Downloads a file from a URL, and write it to a user defined location
@@ -17,6 +17,8 @@ def download_file(url, write_dir):
     Args:
         url (str): download url
         write_dir (str): Host directory to which the data will be written
+        overwrite (bool): Should the file be overwritten if already existing on local host
+        check_integrity (bool): Only makes sense if overwrite is set to False (when updating the archive)
 
     Details:
         The function will raise a ConnectionError (via requests) if a connection cannot be established
@@ -27,16 +29,29 @@ def download_file(url, write_dir):
         str: The filename of the downloaded data
     """
     local_filename = os.path.join(write_dir, url.split('/')[-1])
+    file_exists = os.path.isfile(local_filename)
+    if file_exists and not overwrite and not check_integrity:
+        # No need to even send a request, end function
+        return local_filename           
     # Test that file is present on remote
     r0 = requests.head(url)
     if r0.status_code == 404:
         raise HttpResourceNotAvailable
-    r = requests.get(url, stream=True)
+    if not overwrite and file_exists and check_integrity:
+        # FIle integrity checking
+        if os.path.getsize(local_filename) == int(r0.headers['Content-Length']): # file size matches
+            return local_filename
+    # Create directory if it doesn't exist yet
+    if not os.path.exists(write_dir):
+        os.makedirs(write_dir)
+    # Download file
+    r1 = requests.get(url, stream=True)
     with open(local_filename, 'wb') as f:
-        for chunk in r.iter_content(chunk_size=1024):
+        for chunk in r1.iter_content(chunk_size=1024):
             if chunk:
                 f.write(chunk)
     return local_filename
+
 
 
 def download_to_tree(url, base_dir, overwrite = False, check_integrity = False):
